Fills in the more filter logic in the README, saves some diagram changes I made so long ago I don't remember them, and gets rid of the dummySweeper.
Since the filters don't store data in the DB, we don't need to worry about sweeping for multiple values for one day because of dummy data that was later filled in with real data. We'll get one sample from the data store per day, and if there aren't any for a day the interpolator will add them right before the trend is calculated.