Implements the trend filter and revises (and simplifies) the structure of the filter class. It's not reading and writing to the DB between each filter layer or between calls. So instead of being a standalone operation that's persisted in the DB, filtering happens in memory upon request when data is fetched. The controller just runs the samples through the filter before sending them to the caller.
This means the filter itself returns an array of data samples in its new custom callback, instead of just being an on-completion hook to let the caller know it's safe to grab them again from the data store. And the filter layers no longer have to be async from the data store calls they were going to do, so the pipeline itself doesn't need dispatch and semaphores. The layers' closure signature has changed too, so it's now a simple array of DataSamples in and out.

The trend filter itself is pretty basic. A one-item buffer keeps the last sample around as we loop through, and changes smoothed out over a 10 moving average.

For now the test is just checking that, upon receipt after filtering, the most recent dummy weight's calculated trend rounds to 193.7 lbs. Also, calling purge from the data store test was confusing Xcode because it didn't fulfill expectations over there, generating a bogus test failure. Switched to doing the same work inline.